Summary:
This chapter provided a very brief introduction to:

Entity Framework.

Creating a domain model and using Entity Framework to map that model to a database.

How to change the mapping of the model and seed the database with sample data.

TODO:

Revisit EF5 to adjust the model to support new requirements' implementation.

Domain-Driven Design and to leverage some of its principles to help write more maintainable code as well as the effect of those principles on how to structure the code.

Modify app to adhere to DDD principles.